+++ pkg/controller/add_kogitosupportingservice.go
@@ -20,5 +20,5 @@ import (
 
 func init() {
 	// AddToManagerFuncs is a list of functions to create controllers and add them to a manager.
-	AddToManagerFuncs = append(AddToManagerFuncs, kogitosupportingservice.Add, kogitosupportingservice.AddDataIndex)
+	AddToManagerFuncs = append(AddToManagerFuncs, kogitosupportingservice.Add)
 }