+++ pkg/controller/kogitoruntime/kogitoruntime_controller.go
@@ -112,6 +112,11 @@ func (r *ReconcileKogitoRuntime) Reconcile(request reconcile.Request) (result re
 		return
 	}
 
+	if err = infrastructure.MountProtoBufConfigMapOnDataIndex(r.client, instance); err != nil {
+		log.Errorf("Fail to mount Proto Buf config map of Kogito runtime service(%s) on DataIndex", instance.Name, err)
+		return
+	}
+
 	definition := services.ServiceDefinition{
 		Request:            request,
 		DefaultImageTag:    infrastructure.LatestTag,