+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@ third party infrastructure resources are not yet available
 - [KOGITO-3556](https://issues.redhat.com/browse/KOGITO-3556) - Review CRDs Status resource update
  
 ## Bug Fixes
+- [KOGITO-3736](https://issues.redhat.com/browse/KOGITO-3736) - Reconciler error in Data index controller when KogitoRuntime is deployed
 
 ## Known Issues
 The protobuf ConfigMap does not update in Spring Boot due to [this issue](https://issues.redhat.com/browse/KOGITO-3406).